Rename plugin to BatchFarming
In addition to sowing, functions are being added, so we change the name.
