+++ src/kim/present/batchfarming/event/UseBatchFarmingEvent.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace kim\present\showysowing\event;
+namespace kim\present\batchfarming\event;
 
 use pocketmine\event\Cancellable;
 use pocketmine\event\CancellableTrait;
